+++ feature/setting/src/main/java/team/ppac/setting/navigation/PrivacyPolicyNavigation.kt
@@ -3,7 +3,7 @@ package team.ppac.setting.navigation
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import team.ppac.setting.PrivacyPolicyRoute
+import team.ppac.setting.policy.PrivacyPolicyRoute
 
 const val PRIVACY_POLICY_ROUTE = "privacy_policy"
