+++ feature/setting/src/main/java/team/ppac/setting/policy/PrivacyPolicyScreen.kt
@@ -1,4 +1,4 @@
-package team.ppac.setting
+package team.ppac.setting.policy
 
 import android.view.ViewGroup
 import android.webkit.WebView
